refactor(yargs): move most of yargs to lib/yargs
